New XML Interface:
Now the results can be downloaded in xml files via the server. Example
of such http call to server is:
http://eppic-web.org/ewui/ewui/dataDownload?type=xml&id=1smt,2gs2&getSeqInfo=f

The following fields are required to be provided in the url:
a) type: only possible value is xml for now!
b) id: a list of comma seperated values of either EPPIC-Job Ids or PDB
Codes. Note that PDB codes should be in lower case.

The following field is optional:
getSeqInfo: possible values: t/f; tells whether the xml file should
contain sequence information or not. Default(if not provided) is t!