First version of 3Dmol LatticeGraph
Refactors LatticeGUI into a datastructure (LatticeGraph3D)
to hold the positions, colors, etc., and some thin display
code based on mustache templates for the actual GUI.

A few discrepencies remain between the old LatticeGUI and
the new LatticeGUIJmol, after which the former will be
removed.

Adds a dependency on mustache.java.